Handle changes from PR #31
